+++ tests/continuations.scm
@@ -1,37 +1,56 @@
 (test.failing "continuations: base"
-  (lambda (t)
-    (define x 0)
+      (lambda (t)
+        (define x 0)
 
-    (t.is (+ 2 (call/cc (lambda (cc)
-                           (set! x cc)
-                            3)))
-           5)
+        (t.is (+ 2 (call/cc (lambda (cc)
+                              (set! x cc)
+                              3)))
+              5)
 
-    (t.is (x 4) 6)))
+        (t.is (x 4) 6)))
 
 (test.failing "continuations: return"
-       (lambda (t)
-          (let ((called #f))
+      (lambda (t)
+        (let ((called #f))
 
-            (define (bar)
-              (set! called #t))
+          (define (bar)
+            (set! called #t))
 
-            (define (foo)
-               (call/cc (lambda (return)
-                          (return 10)
-                          (bar))))
+          (define (foo)
+            (call/cc (lambda (return)
+                       (return 10)
+                       (bar))))
 
-            (t.is (foo) 10)
-            (t.is called #f))))
+          (t.is (foo) 10)
+          (t.is called #f))))
 
 (test.failing "continuations: calling"
-       (lambda (t)
-          (let ((called))
-            (t.is (let ((my-val (call/cc (lambda (c) c))))
-                    (if (procedure? my-val)
-                        (my-val 10)
-                        (begin
-                            (set! called #t)
-                            my-val)))
-                  10)
-             (t.is called #t))))
+      (lambda (t)
+        (let ((called))
+          (t.is (let ((my-val (call/cc (lambda (c) c))))
+                  (if (procedure? my-val)
+                      (my-val 10)
+                      (begin
+                        (set! called #t)
+                        my-val)))
+                10)
+          (t.is called #t))))
+
+;; example that found a bug in BiwaScheme
+;; https://github.com/biwascheme/biwascheme/issues/257
+(test.failing "continuations: saving/restoring environment"
+      (lambda (t)
+        (let ((result (call/cc (lambda (return)
+                                 (let ((n 5)
+                                       (result (list))
+                                       (k #f))
+                                   (set! result (append result (list (call/cc (lambda (return)
+                                                                                (set! k return)
+                                                                                "Hello")))))
+                                   (when #t
+                                     (if (zero? n)
+                                         (return result))
+                                     (set! n (- n 1))
+                                     (k (string-append "Hello <" (number->string n) ">"))))))))
+
+          (t.is result '("Hello <0>")))))